simplifying interface of GO enrichment classes
