--- phenol-analysis/src/main/java/org/monarchinitiative/phenol/analysis/PopulationSet.java
@@ -1,15 +1,0 @@
-package org.monarchinitiative.phenol.analysis;
-
-import org.monarchinitiative.phenol.ontology.data.TermId;
-
-import java.util.Map;
-import java.util.Set;
-
-public class PopulationSet extends StudySet {
-
-  public PopulationSet(Set<TermId> genes,
-                       Map<TermId, DirectAndIndirectTermAnnotations> associationContainer) {
-
-    super(genes, "population set", associationContainer);
-  }
-}
