+++ phenol-analysis/src/main/java/org/monarchinitiative/phenol/stats/ParentChildIntersectionPValueCalculation.java
@@ -2,7 +2,7 @@ package org.monarchinitiative.phenol.stats;
 
 
 import com.google.common.collect.Sets;
-import org.monarchinitiative.phenol.analysis.AssociationContainer;
+import org.monarchinitiative.phenol.analysis.TermAssociationContainer;
 import org.monarchinitiative.phenol.analysis.StudySet;
 import org.monarchinitiative.phenol.ontology.data.Ontology;
 import org.monarchinitiative.phenol.ontology.data.TermId;
@@ -20,7 +20,7 @@ import java.util.Set;
  */
 public class ParentChildIntersectionPValueCalculation extends ParentChildPValuesCalculation {
 	public ParentChildIntersectionPValueCalculation(Ontology graph,
-                                                  AssociationContainer goAssociations,
+                                                  TermAssociationContainer goAssociations,
                                                   StudySet populationSet,
                                                   StudySet studySet,
                                                   MultipleTestingCorrection mtc) {
